+++ persistent_store/etcdv3_test.go
@@ -622,6 +622,8 @@ func TestEtcdv3AddSolidFireBackend(t *testing.T) {
 	sfConfig := drivers.SolidfireStorageDriverConfig{
 		CommonStorageDriverConfig: &drivers.CommonStorageDriverConfig{
 			StorageDriverName: drivers.SolidfireSANStorageDriverName,
+		},
+		SolidfireStorageDriverConfigDefaults: drivers.SolidfireStorageDriverConfigDefaults{
 			CommonStorageDriverConfigDefaults: drivers.CommonStorageDriverConfigDefaults{
 				Size: "1GiB",
 			},