+++ storage_drivers/ontap/ontap_nas_qtree.go
@@ -218,9 +218,9 @@ func (d *NASQtreeStorageDriver) Create(name string, sizeBytes uint64, opts map[s
 		return fmt.Errorf("volume %s already exists", name)
 	}
 
-	if sizeBytes < MinimumVolumeSizeBytes {
-		return fmt.Errorf("requested volume size (%d bytes) is too small; the minimum volume size is %d bytes",
-			sizeBytes, MinimumVolumeSizeBytes)
+	sizeBytes, err = GetVolumeSize(sizeBytes, d.Config)
+	if err != nil {
+		return err
 	}
 
 	// Ensure qtree name isn't too long